requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiologger==0.7.0
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.11.1
APScheduler==3.10.1
apscheduler-di==0.1.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
fake-useragent==1.1.1
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isort==5.12.0
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
multidict==6.0.4
packaging==23.0
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
rodi==2.0.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.15
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
yarl==1.8.2
